+++ controllers/organizer-controller.js
@@ -3,30 +3,32 @@ const bodyParser = require('body-parser');
 const connection = require('./../config/db-connection.js');
 
 exports.searchOrganizer = (req, res) => {
-	query = 'CALL search_organizer(?)';
+	query = 'SELECT * FROM organizer where name like ? or description like ?';
 
-	connection.userType('A').query(query, 
-		[
-			"%" + req.query.search + "%"
-		], function (err, rows) {
-			if(!err) {
-				if(rows[0].length == 1) {
-					res.status(200).send(rows[0][0]);
-					return rows[0][0];
-				} else {
-					res.status(200).send(rows[0]);
-					return rows;
-				}
+	connection.query(query, ["%" + req.query.search + "%", "%" + req.query.search + "%"], function (err, rows) {
+		if(!err) {
+			if(rows.length == 1) {
+				res.status(200).send(rows[0]);
+				return rows[0];
 			} else {
-				res.status(500).send({'message' : 'Internal Server Error'});
+				res.status(200).send(rows);
+				return rows;
 			}
+		} else {
+			res.status(500).send({'message' : 'Internal Server Error'});
+		}
 	});
 }
 
 exports.editOrganizer = (req,res) => {
 	currentUser = req.session.user;
 
-	connection.query('UPDATE organizer SET name = ?, description = ? WHERE id = ?', [req.body.name, req.body.description, req.body.id], function(err, rows){
+	connection.query('UPDATE organizer SET name = ?, description = ? WHERE id = ?', 
+		[
+			req.body.name, 
+			req.body.description, 
+			req.body.id
+		], function(err, rows){
 		if(!err) {
 			connection.query('SELECT * from organizer where id = ?', [req.body.id], function(err, rows) {
 				if(!err) {
@@ -42,10 +44,11 @@ exports.editOrganizer = (req,res) => {
 }
 
 exports.findGames = (req,res,next) =>{
-
-	connection.query("SELECT * from game WHERE game.organizer_id = ?"
-		, [req.query.id],
-		function(err,rows){
+	query = "CALL findGames"
+	connection.userType('A').query(query, 
+		[
+			req.query.id
+		], function(err,rows){
 		if(!err){
 			if(row.length == 1){
 				res.status(200).send({'message' : 'Sucessfully Retrieved Info'},rows[0]);
@@ -55,16 +58,17 @@ exports.findGames = (req,res,next) =>{
 				return rows;
 			}
 		} else {
-			res.status(500).send({'message' : 'Internal Server Error'})
+		 	res.status(500).send({'message' : 'Internal Server Error'})
 		}
 		});
-};
-
+};		 
+		 
 exports.findSport = (req,res,next) =>{
-
-	connection.query("SELECT * from sport WHERE sport.game_id = ?"
-		, [req.query.game_id],
-		function(err,rows){
+	query = "CALL findSport(?)"
+	connection.userType('A').query(query, 
+		[
+		 	req.query.game_id
+		], function(err,rows){
 		if(!err){
 			if(row.length == 1){
 				res.status(200).send({'message' : 'Sucessfully Retrieved Info'},rows[0]);
@@ -80,10 +84,11 @@ exports.findSport = (req,res,next) =>{
 };
 
 exports.findTeam = (req,res,next) =>{
-
-	connection.query("SELECT * from team WHERE team.sport_id = ?"
-		, [req.query.sport_id],
-		function(err,rows){
+	query = "CALL findTeam(?)"
+	connection.userType('A').query(query, 
+		[
+			req.query.sport_id
+		], function(err,rows){
 		if(!err){
 			if(row.length == 1){
 				res.status(200).send({'message' : 'Sucessfully Retrieved Info'},rows[0]);
@@ -100,13 +105,15 @@ exports.findTeam = (req,res,next) =>{
 
 
 exports.getRequest = (req, res, next) => {
-	connection.query("SELECT * from team WHERE team_id = ?"
-	, [req.query.team_id],
-		function(err,rows){
+	query = "CALL getRequest(?)"
+	connection.userType('A').query(query, 
+		[
+			req.query.team_id
+		], function(err,rows){
 		if(!err){
 			if(row.length == 1){
 				res.status(200).send({'message' : 'Sucessfully Retrieved Info'},rows[0]);
-				return rows[0];
+				return(rows[0][0]);
 			} else {
 				res.status(200).send({'message' : 'Sucessfully Retrieved Info'},rows);
 				return rows;
@@ -119,10 +126,12 @@ exports.getRequest = (req, res, next) => {
 
 
 exports.acceptRequest = (req, res, next) => {
-
-	connection.query("UPDATE team SET pending_participation = TRUE WHERE team_id =?"
-	, [req.query.team_id],
-	function(err,rows){
+	query = "CALL acceptRequest(?)"
+	console.log(query);
+	connection.userType('A').query(query, 
+		[
+			req.query.team_id
+		], function(err,rows){
 		if(!err){
 			res.status(200).send({'message' : 'Sucuessfully Updated Request'});
 			connection.query("SELECT * from team WHERE team_id = ? "
