--- ringo/config/organizer-procedures.sql
@@ -1,86 +1,0 @@
-USE malicsi;
-DROP procedure IF EXISTS rankings;
-DELIMITER //
-
-CREATE PROCEDURE findGame ( IN userID INT) 
-
-
-BEGIN
-
-
-   SELECT * from game WHERE game.organizer_id = userID;
-
-
-END; //
-
-DELIMITER ;
-
-DROP procedure IF EXISTS rankings;
-DELIMITER //
-
-CREATE PROCEDURE findSport ( IN gameID INT) 
-
-
-BEGIN
-
-
-   SELECT * from sport WHERE sport.game_id = gameID;
-
-
-END; //
-
-DELIMITER ;
-
-DROP procedure IF EXISTS rankings;
-DELIMITER //
-
-CREATE PROCEDURE findTeam ( IN teamID INT) 
-
-
-BEGIN
-
-
-   SELECT * from team WHERE team.sport_id = teamID;
-
-
-END; //
-
-DELIMITER ;
-
-DROP procedure IF EXISTS rankings;
-DELIMITER //
-
-CREATE PROCEDURE getRequest ( IN teamID INT) 
-
-
-BEGIN
-
-
-   SELECT * from team WHERE team_id = teamID;
-
-
-END; //
-
-DELIMITER ;
-
-DROP procedure IF EXISTS rankings;
-DELIMITER //
-
-CREATE PROCEDURE acceptRequest ( IN teamID INT) 
-
-
-BEGIN
-
-
-   UPDATE team SET pending_participation = TRUE WHERE team_id =teamID;
-
-
-END; //
-
-DELIMITER ;
-
-grant execute on procedure findGame to organizer
-grant execute on procedure findSport to organizer
-grant execute on procedure findTeam to organizer
-grant execute on procedure getRequest to organizer
-grant execute on procedure acceptRequest to organizer
